Rename stackdriver -> google cloud where applicable
