+++ exporter/googlecloudexporter/go.mod
@@ -1,4 +1,4 @@
-module github.com/open-telemetry/opentelemetry-collector-contrib/exporter/stackdriverexporter
+module github.com/open-telemetry/opentelemetry-collector-contrib/exporter/googlecloudexporter
 
 go 1.14
