+++ exporter/googlecloudexporter/resource_mapper.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package stackdriverexporter
+package googlecloudexporter
 
 import (
 	"contrib.go.opencensus.io/exporter/stackdriver"
